* **26.08.2019** - **implementing different rarefaction depths analysis**   * possible scripts to **modify** are:     * `/Users/paul/Documents/CU_combined/Github/115_isolate_taxa.sh` - **ok**     * `/Users/paul/Documents/CU_combined/Github/130_get_core_metrics_non_phylogenetic.sh` - **ok**     * `/Users/paul/Documents/CU_combined/Github/131_get_core_metrics_non_phylogenetic_collpased.sh` - **ok**     * and more (R scripts, mantel and procrustes tests - after Mandanas input?) - **see below**   * adjusting `/Users/paul/Documents/CU_combined/Github/115_isolate_taxa.sh`     * namely `string[2]='Eukaryote-shallow'` and loop counters     * running script, ok, created files: `/Users/paul/Documents/CU_combined/Zenodo/Qiime/115_*_Eukaryote-shallow.qza`     * commit `d52f11e7a706dac928122533ed6b92a09b95131a` and later `f3b8a5bea7a7bf6cd3bba65f54787832546e87ad`   * adjusted hostnames in work scripts: `find /Users/paul/Documents/CU_combined/Github -name '*.sh' -exec gsed -i 's|"pc683.eeb.cornell.edu"|"macmini.staff.uod.otago.ac.nz"|g' {} \;`     * commit   * adjusting `/Users/paul/Documents/CU_combined/Github/130_get_core_metrics_non_phylogenetic.sh`     * checking rarefaction curve 1 `qiime tools view /Users/paul/Documents/CU_combined/Zenodo/Qiime/125_18S_eDNA_samples_tab_Eukaryotes_non_phylogenetic_curves.qzv` - **aborted**     * checking rarefaction curve 2 `qiime tools view /Users/paul/Documents/CU_combined/Zenodo/Qiime/160_eDNA_samples_Eukaryotes_curves_tree-matched.qzv` - ok     * setting shallow depth to 40000 sequences - ok     * done via new case - run script - see folder `130_18S_eDNA_samples_Eukaryote-shallow_core_metrics_non_phylogenetic` - ok
```
      *"Eukaryote-shallow"* )
      depth=40000
      echo "${bold}Depth set to $depth for Eukaryotes (shallow set)...${normal}"
      ;;
```
  * adjusting script `/Users/paul/Documents/CU_combined/Github/131_get_core_metrics_non_phylogenetic_collpased.sh`
    * inserted new `case` statement
    * running - ok
  * possible scripts to **modify** and  **re-run** are:
    * `/Users/paul/Documents/CU_combined/Github/140_seq_align_mask.sh` - ok
    * `/Users/paul/Documents/CU_combined/Github/145_alignment_export.sh` - **modified**

+++ 135_seq_align.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# 31.05.2019 - Paul Czechowski - paul.czechowski@gmail.com 
+# 26.08.2019 - Paul Czechowski - paul.czechowski@gmail.com 
 # ========================================================
 # Visualising reads after denoising and merging procedure.
 
@@ -98,11 +98,22 @@ done
 
 for k in "${!inpth_seq[@]}"; do
   
-  printf "\n${bold}$(date):${normal} Alignining file ${inpth_seq[$k]}...\n"
-  qiime alignment mafft \
-    --i-sequences "${inpth_seq[$k]}" \
-    --o-alignment "${otpth_seq[$k]}" \
-    --p-n-threads "$thrds" \
-    --verbose 2>&1 | tee -a "${otpth_log[$k]}"
+  # actual filtering
+  # continue only if output file isn't already there
+  if [ ! -f "${otpth_seq[$k]}" ]; then
+  
+    printf "\n${bold}$(date):${normal} Alignining file ${inpth_seq[$k]}...\n"
+    qiime alignment mafft \
+      --i-sequences "${inpth_seq[$k]}" \
+      --o-alignment "${otpth_seq[$k]}" \
+      --p-n-threads "$thrds" \
+      --verbose 2>&1 | tee -a "${otpth_log[$k]}"
+  
+  else
+ 
+    # diagnostic message
+    printf "${bold}$(date):${normal} Analysis already done for \"$(basename "${inpth_seq[$k]}")\"...\n"
+
+  fi
     
 done
